requirements.txt
# FastAPI - Modern web framework for building APIs
fastapi==0.104.1

# Uvicorn - ASGI server for running FastAPI
uvicorn[standard]==0.24.0

# SQLAlchemy - ORM for database operations
sqlalchemy==2.0.23

# Pydantic - Data validation using Python type annotations
pydantic==2.5.0
pydantic-settings==2.1.0

# Python-multipart - For form data parsing
python-multipart==0.0.6

# Database driver for SQLite
aiosqlite==0.19.0

# Logging
python-json-logger==2.0.7

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-playwright==0.4.3  # Playwright integration with pytest
httpx==0.25.1  # Required by TestClient